Fix dynamic resizing of menu grid view
Fix (mostly) frequent layout of menu grid view columns when device (iPad)
reorients slightly. This solution constrains the POSView to resize the
MenuGridView and OrderView widths only when changing to/from Landscape
and Portrait orientations.

The "mostly" caveat is that occasionally in landscape mode the menu grid
will shift to-from 3-4 columns for unknown reasons, but with this fix
it is no longer frequent and jittery.