KDS feature
This feature displays orders with .inProgress and .processed status in a grid view.
Tapping on an order in KDS view will "advance" the status, i.e. from .inProcess to
.processed and from .processed to .delivered, the latter no longer displaying.

Incomplete orders, i.e. unpaid, are "recycled" when the app launches for display in
the POS order view,  which avoids orphaned orders accumulating in Big Peer.

- POS grid view layout anomalies when rotating
- text breaking in POS order view on small devices

- some additional renaming refactoring for KDS-specific objects
- KDS order views in a grid view, which advance order status when tapped
- unpaid order recovery at launch
- move selectedTab from POS_VM to MainVM