Eviction feature with incomplete outgoing order reset
Outgoing order reset means clearing all order items from an unpaid order
when changing location on that device. This is to prevent orders begun
on one device from lingering incomplete in the KDS view of another device
when changing location.

- eviction feature based on compile-time constant OrderTTL set at 24hrs:
When the app goes to background orders older than OrderTTL are evicted
- orders subscription updated to filter orders older than OrderTTL
- POS_VM implements a notification listener that fires before published
dittoService.locationId is set in the Locations view. If a current order
is incomplete (unpaid) when the location changes, it will be reset
- comment and purge print statements